# Optimized FromS7Conversions for arrays