--- a/Sally7.Tests/ArrayConversionTests.cs
+++ b/Sally7.Tests/ArrayConversionTests.cs
@@ -8,8 +8,27 @@ namespace Sally7.Tests
     public class ArrayConversionTests
     {
         [Theory]
-        [InlineData(new int[] {0})]
-        [InlineData(new int[] {1, 1 << 8, 1 << 16, 1 << 24})]
+        [InlineData(new short[] { 1 })]
+        [InlineData(new short[] { 1, 1 << 8 })]
+        public void ConvertToShortArray(short[] value)
+        {
+            var bytes = value.SelectMany(v =>
+            {
+                var b = BitConverter.GetBytes(v);
+                if (BitConverter.IsLittleEndian) Array.Reverse(b);
+                return b;
+            }).ToArray();
+
+            var converter = ConverterFactory.GetFromPlcConverter<short[]>();
+            var result = new short[value.Length];
+            converter(ref result, bytes, value.Length);
+
+            Assert.Equal(value, result);
+        }
+
+        [Theory]
+        [InlineData(new int[] { 1 })]
+        [InlineData(new int[] { 1, 1 << 8, 1 << 16, 1 << 24 })]
         public void ConvertToIntArray(int[] value)
         {
             var bytes = value.SelectMany(v =>
@@ -27,9 +46,9 @@ namespace Sally7.Tests
         }
 
         [Theory]
-        [InlineData(new short[] {0})]
-        [InlineData(new short[] {1, 1 << 8})]
-        public void ConvertToShortArray(short[] value)
+        [InlineData(new long[] { 1 })]
+        [InlineData(new long[] { 1, 1 << 8, 1 << 16, 1 << 24, 1 << 32, 1 << 40, 1 << 48, 1 << 56 })]
+        public void ConvertToLongArray(long[] value)
         {
             var bytes = value.SelectMany(v =>
             {
@@ -38,8 +57,46 @@ namespace Sally7.Tests
                 return b;
             }).ToArray();
 
-            var converter = ConverterFactory.GetFromPlcConverter<short[]>();
-            var result = new short[value.Length];
+            var converter = ConverterFactory.GetFromPlcConverter<long[]>();
+            var result = new long[value.Length];
+            converter(ref result, bytes, value.Length);
+
+            Assert.Equal(value, result);
+        }
+
+        [Theory]
+        [InlineData(new float[] { 3.14f })]
+        [InlineData(new float[] { 2.81f, 3.14f })]
+        public void ConvertToFloatArray(float[] value)
+        {
+            var bytes = value.SelectMany(v =>
+            {
+                var b = BitConverter.GetBytes(v);
+                if (BitConverter.IsLittleEndian) Array.Reverse(b);
+                return b;
+            }).ToArray();
+
+            var converter = ConverterFactory.GetFromPlcConverter<float[]>();
+            var result = new float[value.Length];
+            converter(ref result, bytes, value.Length);
+
+            Assert.Equal(value, result);
+        }
+
+        [Theory]
+        [InlineData(new double[] { 3.14 })]
+        [InlineData(new double[] { 2.81, 3.14 })]
+        public void ConvertToDoubleArray(double[] value)
+        {
+            var bytes = value.SelectMany(v =>
+            {
+                var b = BitConverter.GetBytes(v);
+                if (BitConverter.IsLittleEndian) Array.Reverse(b);
+                return b;
+            }).ToArray();
+
+            var converter = ConverterFactory.GetFromPlcConverter<double[]>();
+            var result = new double[value.Length];
             converter(ref result, bytes, value.Length);
 
             Assert.Equal(value, result);